import errors fixed + utils no longer import Notebook/NotebookPage circularity fix
