+++ coppafish/utils/__init__.py
@@ -1,3 +1,1 @@
 from .base import reed_solomon_codes, estimate_runtime
-from .pciseq import export_to_pciseq
-from .version import CompatibilityTracker
